add lvalue ref variant matcher
