+++ include/refl-cpp/impl/variant_matcher.hpp
@@ -1,77 +1,15 @@
 #pragma once
 
-#include "../variant.hpp"
+#include "refl-cpp/variant.hpp"
 
-namespace ReflCpp::detail {
-template <VariantWrapperType Type, typename R>
-struct VariantMatcher {
-    static bool Match(const TypeID _) {
-#ifndef NDEBUG
-        throw std::logic_error("Not implemented\n" + to_string(std::stacktrace::current()));
-#else
-        return false;
-#endif
-    }
-};
+#include "refl-cpp/impl/variant_matcher/value_variant_matcher.hpp"
+#include "refl-cpp/impl/variant_matcher/const_value_variant_matcher.hpp"
+#include "refl-cpp/impl/variant_matcher/lvalue_ref_variant_matcher.hpp"
 
+// There is technically no need for this since we just default to 'false' anyway
 template <typename R_>
-struct VariantMatcher<VariantWrapperType::VOID, R_> {
-    static bool Match(const TypeID _) {
+struct ReflCpp::detail::VariantMatcher<ReflCpp::detail::VariantWrapperType::VOID, R_> {
+    static bool Match(const TypeID) {
         return false;
     }
 };
-
-template <typename R_>
-    requires (!std::is_pointer_v<R_> && !std::is_reference_v<R_>)
-struct VariantMatcher<VariantWrapperType::VALUE, R_> {
-    using ResultT = R_&;
-
-    static bool Match(const TypeID type) {
-        return type.Equals<R_>();
-    }
-
-    static ResultT Get(VariantBase* base) {
-        return static_cast<VariantWrapper<ResultT>*>(base)->GetValue();
-    }
-};
-
-template <typename R_>
-    requires (!std::is_pointer_v<R_> && !std::is_reference_v<R_>)
-struct VariantMatcher<VariantWrapperType::VALUE, const R_> {
-    using ResultT = R_&;
-
-    static bool Match(const TypeID type) {
-        return type.Equals<R_>()
-            || type.Equals<const R_>();
-    }
-
-    static ResultT Get(VariantBase* base) {
-        return static_cast<VariantWrapper<ResultT>*>(base)->GetValue();
-    }
-};
-
-template <typename R_>
-struct VariantMatcher<VariantWrapperType::VALUE, R_&> {
-    using ResultT = R_&;
-
-    static bool Match(const TypeID type) {
-        return type.Equals<R_>();
-    }
-
-    static ResultT Get(VariantBase* base) {
-        return static_cast<VariantWrapper<ResultT>*>(base)->GetValue();
-    }
-};
-
-template <typename R_>
-struct VariantMatcher<VariantWrapperType::VALUE, const R_&> {
-    static bool Match(const TypeID type) {
-        return type.Equals<R_>()
-            || type.Equals<const R_>();
-    }
-
-    static const R_& Get(VariantBase* base) {
-        return static_cast<VariantWrapper<const R_&>*>(base)->GetValue();
-    }
-};
-}
